requirements.txt
boto3==1.9.167
botocore==1.12.167
cairocffi==1.0.2
CairoSVG==2.4.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-appconf==1.0.3
django-disqus==0.5
django-mathfilters==0.4.0
django-money==0.15
django-storages==1.7.1
django-tagging==0.4.6
docutils==0.14
html5lib==1.0.1
idna==2.8
jmespath==0.9.4
mock==3.0.5
mysqlclient==1.4.2
oauthlib==3.0.1
pilkit==2.0
Pillow==6.0.0
pop==5.2.0
py-moneyed==0.8.0
pycparser==2.19
PyMySQL==0.9.3
Pyphen==0.9.5
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
tinycss2==1.0.2
urllib3==1.25.3
webencodings==0.5.1
